Fix warnings about incorrect integer types
Mostly by changing int to long/NSInteger/NSUInteger

Add long_to_integer_bound() function to explicitly
convert from long to the required field length in
PBGitGrapher.mm as discussed in #16.

+++ Classes/Controllers/ApplicationController.m
@@ -62,7 +62,7 @@ static OpenRecentController* recentsDialog = nil;
 	[NSApp setServicesProvider:services];
 
 	// Force update the services menu if we have a new services version
-	int serviceVersion = [[NSUserDefaults standardUserDefaults] integerForKey:@"Services Version"];
+	NSInteger serviceVersion = [[NSUserDefaults standardUserDefaults] integerForKey:@"Services Version"];
 	if (serviceVersion < 2)
 	{
 		NSLog(@"Updating services menu…");
