+++ Classes/Views/PBQLOutlineView.m
@@ -67,7 +67,7 @@
 
 		// select the row that was clicked before showing the menu for the event
 		NSPoint mousePoint = [self convertPoint:[theEvent locationInWindow] fromView:nil];
-		int row = [self rowAtPoint:mousePoint];
+		NSInteger row = [self rowAtPoint:mousePoint];
 
 		// figure out if the row that was just clicked on is currently selected
 		if ([selectedRowIndexes containsIndex:row] == NO) {
